+++ examples/type_algebra/main.cpp
@@ -380,8 +380,8 @@ auto test_identity_transformation(fn::just<UserId> j) -> void
 // sync-example-test-choice-mapping
 auto test_choice_mapping(fn::choice<User, UserId> ch) -> void
 {
-  constexpr auto mapper = fn::overload{[](UserId) { return fn::choice<Missing>{Missing{}}; },
-                                       [](User) { return fn::choice<FilePath>{FilePath{}}; }};
+  constexpr auto mapper = fn::overload{[](UserId) { return fn::as_choice(Missing{}); },
+                                       [](User) { return fn::as_choice(FilePath{}); }};
 
   // transform nests the returned choice as a mapped value
   auto mapped = ch | fn::transform(mapper);